refactor(frontend): move toast/drawer/modal/escapeHtml to js/ui.js and import in app.js
